requirements.txt
blinker==1.3
certifi==2020.4.5.2
chardet==3.0.4
cloud-init==18.4
command-not-found==0.3
configobj==5.0.6
crcmod==1.7
cryptography==1.2.3
decorator==4.0.6
idna==2.9
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
markovify==0.8.0
MarkupSafe==0.23
numpy==1.18.5
oauthlib==3.1.0
pandas==0.24.2
Pillow==7.2.0
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
PySocks==1.7.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.1
python-debian==0.1.27
python-dotenv==0.13.0
python-systemd==231
pytz==2020.1
PyYAML==3.11
requests==2.23.0
requests-oauthlib==1.3.0
scipy==0.17.0
six==1.15.0
ssh-import-id==5.5
tweepy==3.8.0
ufw==0.35
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.9
